Add some adjusts
