+++ CinemaTV/TV Shows/Views/TVShowView.swift
@@ -45,3 +45,7 @@ struct TVShowView: View {
         }
     }
 }
+
+#Preview {
+    TVShowView()
+}
